+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.4
 * Add support for custom cache serialization
+* Serialize response header in cache documents
 
 ## 0.0.3
 * Rename utility cache functions and fix functionality